requirements.txt
google-generativeai
gradio
python-pptx
pandas
openpyxl
google-generativeai
gradio
python-pptx
pandas
openpyxl
google-generativeai
gradio
python-pptx
pandas
openpyxl
requests